First implementation of AsyncFirecrest client
