+++ firecrest/__init__.py
@@ -19,12 +19,9 @@ if sys.version_info[:3] < MIN_PYTHON_VERSION:
     )
     sys.exit(1)
 
-from firecrest.BasicClient import (
-    Firecrest,
-    ExternalDownload,
-    ExternalUpload,
-    ExternalStorage,
-)
+from firecrest.BasicClient import Firecrest
+from firecrest.AsyncClient import AsyncFirecrest
+from firecrest.ExternalStorage import ExternalDownload, ExternalUpload, ExternalStorage
 from firecrest.Authorization import ClientCredentialsAuth
 from firecrest.FirecrestException import (
     ClientsCredentialsException,
